requirements.txt

coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.2.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
Jinja2==3.1.2
requests==2.28.2
uritemplate==4.1.1
